+++ include/fb/game/item/cash.h
@@ -8,21 +8,21 @@ namespace fb::game {
 class cash : public item
 {
 public:
-    const uint32_t value = 0;
+    const uint64_t value = 0;
 
 public:
     // clang-format off
-    cash(fb::game::server& server, uint32_t chunk);
+    cash(fb::game::server& server, uint64_t chunk);
    ~cash();
     // clang-format on
 
 private:
-    static const fb::model::cash& match_model(fb::game::server& server, uint32_t value);
+    static const fb::model::cash& match_model(fb::game::server& server, uint64_t value);
 
 public:
     // clang-format off
-    [[nodiscard]] async::task<std::shared_ptr<fb::game::cash>> replace(uint32_t value);
-    [[nodiscard]] async::task<uint32_t>                        reduce(uint32_t value);
+    [[nodiscard]] async::task<std::shared_ptr<fb::game::cash>> replace(uint64_t value);
+    [[nodiscard]] async::task<uint64_t>                        reduce(uint64_t value);
     // clang-format on
 
     // clang-format off
